+++ src/types/index.ts
@@ -1,3 +1,1 @@
-export type Point = [number, number]
-export type Polygon = Point[]
 export type Vec2 = number[] | [number, number] | Float32Array